+++ lib/mongoid/geospatial/helpers/spatial.rb
@@ -10,7 +10,6 @@ Mongoid::Fields.option :spatial do |model, field, _options|
 
   model.class_eval do
     spatial_fields << field.name.to_sym
-    spatial_fields_indexed << field.name.to_sym
 
     # Create 2D index
     spatial_index field.name