+++ lib/mongoid/geospatial/helpers/sphere.rb
@@ -10,7 +10,6 @@ Mongoid::Fields.option :sphere do |model, field, _options|
 
   model.class_eval do
     spatial_fields << field.name.to_sym
-    spatial_fields_indexed << field.name.to_sym
 
     # Create 2Dsphere index
     spherical_index field.name